refactor: register continuation strategies
